requirements.txt
tensorflow
matplotlib
python-chess